Support movements with note removals
Previously this was implict when a movement shrank an interval to 0. Now
we explicitly add :remove as a chord component and rework the core fns
to accommodate for this. The print fns are also reworked, including
changing the alignment of the voicing vectors to maintain per-note
alignment down columns, instead of shifting to align the root. This
makes it much easier to see how the degree of different notes change.